refactor: rename to register to register pass
