+++ Src/Fitnet.IntegrationTests/Passes/RegisterPass/RegisterPassTests.cs
@@ -1,8 +1,7 @@
-namespace SuperSimpleArchitecture.Fitnet.IntegrationTests.Passes.Register;
+namespace SuperSimpleArchitecture.Fitnet.IntegrationTests.Passes.RegisterPass;
 
 using Common.TestEngine;
 using Common.TestEngine.Configuration;
-using Fitnet.Passes;
 using Fitnet.Passes.RegisterPass;
 using Requests;
