+++ Src/Fitnet/Passes/PassesEndpoints.cs
@@ -5,5 +5,5 @@ using RegisterPass;
 internal static class PassesEndpoints
 {
     public static void MapPasses(this IEndpointRouteBuilder app) => 
-        app.MapRegister();
+        app.MapRegisterPass();
 }
